AndABirdSang/requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiohttp-socks==0.7.1
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
cchardet==2.1.7
celery==5.2.7
celery-progress==0.1.3
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dataclasses==0.6
Deprecated==1.2.13
Django==4.1
djangorestframework==3.13.1
elastic-transport==8.4.0
elasticsearch==8.4.3
fake-useragent==0.1.11
frozenlist==1.3.1
geographiclib==1.52
geopy==2.2.0
googletransx==2.4.2
gunicorn==20.1.0
idna==3.3
kombu==5.2.4
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.0
packaging==21.3
pandas==1.5.1
prompt-toolkit==3.0.31
psycopg2-binary==2.9.5
pycares==4.2.2
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-socks==2.0.3
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
schedule==1.1.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tweepy==4.10.1
twint==2.1.21
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
